+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Add support for intervals in seconds
 * Add support for widget opacity
+* Add support for widget offset
 
 ### Changed